requirements.txt
asgiref==3.8.1
Django==5.0.6
django-allauth==0.62.1
django-extensions==3.2.3
googlemaps==4.10.0
psycopg==3.1.18
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
